feat: 【告警中心】TAPD 交互优化与取消授权重构 --story=1135825673
本次改动：
- 修复创建 TAPD 菜单 Popover 未自动关闭的问题
- 取消授权逻辑重构并添加成功提示
- 创建 TAPD 单据时不再关闭 Issue 详情侧栏
- 移除 TAPD 单据类型中的「任务」选项
