+++ bkmonitor/webpack/src/trace/pages/alarm-center/alarm-issues/constant.ts
@@ -360,5 +360,5 @@ export const TAPDWorkspaceBoundEnum = {
 export const TapdTypeMap = [
   { label: window.i18n.t('需求'), value: TapdTypeEnum.STORY },
   { label: window.i18n.t('缺陷'), value: TapdTypeEnum.BUG },
-  { label: window.i18n.t('任务'), value: TapdTypeEnum.TASK },
+  // { label: window.i18n.t('任务'), value: TapdTypeEnum.TASK },
 ];
