+++ bkmonitor/webpack/src/trace/pages/alarm-center/alarm-issues/issues-detail/components/issues-slider-header.scss
@@ -202,7 +202,7 @@
   }
 }
 
-.create-tapd-menu-popover.bk-popover {
+.create-tapd-menu-popover.bk-popover.bk-pop2-content {
   padding: 4px 0;
 
   .create-tapd-menu-popover-content {
