+++ bkmonitor/webpack/src/trace/pages/alarm-center/alarm-issues/issues-detail/issues-detail-sideslider.tsx
@@ -183,7 +183,6 @@ export default defineComponent({
 
     const handleCreateTapd = () => {
       emit('createTapd', true);
-      handleShowChange(false);
     };
 
     const handleConditionChange = (val: IWhereItem[]) => {
